backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
anthropic==0.28.0
apify-client==1.7.1
sqlalchemy==2.0.30
aiosmtplib==3.0.1
email-validator==2.1.1
httpx==0.27.0
pydantic==2.7.1
